Implement better tie checking.
Instead of explicitly checking for ties, we now just prune off parts of
the tree which can not lead to a win. By traversing the tree as the game
is being played, we can see when we have reached a leaf, and that is
when the tie is certain. An interesting consequence of this is that the
basic game logic now requires the presence of the game tree.

This reduces execution time and memory usage slightly, most likely due
to not making explicit tie checks and slightly reduced tree complexity
(26442 vs 26830).